renaming config.rs to auth.rs
