+++ helix-cli/src/commands/auth.rs
